% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{Cap}
\alias{Cap}
\title{Capitalize the first letter}
\usage{
Cap(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string whose first letter is capitalized
}
\description{
Capitalize the first letter
}
\examples{

Cap("michelia")


}
