\name{dat_all_accepted_sp2023}
\alias{dat_all_accepted_sp2023}
\docType{data}
\title{
The accepted names of Checklist of plant species in China (2023 Edition)
}
\description{
More details, see https://www.plantplus.cn/doi/10.12282/plantdata.1390
}
\usage{data("dat_all_accepted_sp2023")}
\format{
  A data frame with 46229 observations on the following 39 variables.
  \describe{
    \item{\code{SP_ID}}{a numeric vector}
    \item{\code{accepted_name_code}}{a character vector}
    \item{\code{name_code}}{a character vector}
    \item{\code{web_site}}{a numeric vector}
    \item{\code{canonical_name}}{a character vector}
    \item{\code{kingdom}}{a character vector}
    \item{\code{kingdom_c}}{a character vector}
    \item{\code{kingdom_c_py}}{a character vector}
    \item{\code{phylum}}{a character vector}
    \item{\code{phylum_c}}{a character vector}
    \item{\code{phylum_c_py}}{a character vector}
    \item{\code{class}}{a character vector}
    \item{\code{class_c}}{a character vector}
    \item{\code{class_c_py}}{a character vector}
    \item{\code{order_c}}{a character vector}
    \item{\code{order_c_py}}{a character vector}
    \item{\code{order}}{a character vector}
    \item{\code{family}}{a character vector}
    \item{\code{family_c}}{a character vector}
    \item{\code{family_c_py}}{a character vector}
    \item{\code{genus}}{a character vector}
    \item{\code{genus_c}}{a character vector}
    \item{\code{genus_c_py}}{a character vector}
    \item{\code{species}}{a character vector}
    \item{\code{species_c}}{a character vector}
    \item{\code{species_c_py}}{a character vector}
    \item{\code{infraspecies_marker}}{a character vector}
    \item{\code{infraspecies}}{a character vector}
    \item{\code{infraspecies_c}}{a character vector}
    \item{\code{infraspecies_c_py}}{a character vector}
    \item{\code{author}}{a character vector}
    \item{\code{year}}{a character vector}
    \item{\code{reference}}{a character vector}
    \item{\code{distribution_c}}{a character vector}
    \item{\code{comment}}{a numeric vector}
    \item{\code{scrutiny_date}}{a character vector}
    \item{\code{is_accepted_name}}{a numeric vector}
    \item{\code{genuswithsubgenus}}{a character vector}
    \item{\code{family_id}}{a character vector}
  }
}
\details{
The checklist of plant species in China is updated annually.
}
\source{
https://www.plantplus.cn/doi/10.12282/plantdata.1390
}
\references{
Checklist of plant species in China (2023 Edition), 2023, Plant Data Center of Chinese Academy of Sciences, DOI:10.12282/plantdata.1390, CSTR:34735.11.PLANTDATA.1390
}
\examples{
data(dat_all_accepted_sp2023)

}
\keyword{datasets}
