\name{dat_all_sp2023}
\alias{dat_all_sp2023}
\docType{data}
\title{
All scientific names included in Checklist of plant species in China (2023 Edition)
}
\description{
More information https://www.plantplus.cn/doi/10.12282/plantdata.1390
}
\usage{data("dat_all_sp2023")}
\format{
  A data frame with 117986 observations on the following 32 variables.
  \describe{
    \item{\code{SP_ID}}{a numeric vector}
    \item{\code{accepted_name_code}}{a character vector}
    \item{\code{name_code}}{a character vector}
    \item{\code{web_site}}{a numeric vector}
    \item{\code{canonical_name}}{a character vector}
    \item{\code{genus}}{a character vector}
    \item{\code{genus_c}}{a character vector}
    \item{\code{species}}{a character vector}
    \item{\code{species_c}}{a character vector}
    \item{\code{infraspecies_marker}}{a character vector}
    \item{\code{infraspecies}}{a character vector}
    \item{\code{infraspecies2}}{a numeric vector}
    \item{\code{infraspecies_c}}{a character vector}
    \item{\code{infraspecies_c2}}{a numeric vector}
    \item{\code{infraspecies_marker2}}{a numeric vector}
    \item{\code{author}}{a character vector}
    \item{\code{author2}}{a numeric vector}
    \item{\code{comment}}{a numeric vector}
    \item{\code{scrutiny_date}}{a character vector}
    \item{\code{sp2000_status_id}}{a numeric vector}
    \item{\code{database_id}}{a character vector}
    \item{\code{specialist_id}}{a numeric vector}
    \item{\code{family_id}}{a character vector}
    \item{\code{is_accepted_name}}{a numeric vector}
    \item{\code{genus_c_py}}{a character vector}
    \item{\code{species_c_py}}{a character vector}
    \item{\code{infraspecies_c_py}}{a character vector}
    \item{\code{infraspecies_c_py2}}{a numeric vector}
    \item{\code{genuswithsubgenus}}{a character vector}
    \item{\code{specialist_code}}{a numeric vector}
    \item{\code{family_code}}{a numeric vector}
    \item{\code{comments}}{a numeric vector}
  }
}
\details{
The checklist of plant species in China is updated annually.
}
\source{
https://www.plantplus.cn/doi/10.12282/plantdata.1390
}
\references{
Checklist of plant species in China (2023 Edition), 2023, Plant Data Center of Chinese Academy of Sciences, DOI:10.12282/plantdata.1390, CSTR:34735.11.PLANTDATA.1390
}
\examples{
data(dat_all_sp2023)
}
\keyword{datasets}
