\name{dat_CBRL2020_higher_plants}
\alias{dat_CBRL2020_higher_plants}
\docType{data}
\title{
Entire Dataset of China Biodiversity Red List (2020) higher plants
}
\description{

Dataset of China Biodiversity Red List (2020) higher plants

}
\usage{data("dat_CBRL2020_higher_plants")}
\format{
  A data frame with 39358 observations on the following 9 variables.
  \describe{
    \item{\code{No}}{a character vector}
    \item{\code{family_cn}}{a character vector}
    \item{\code{family}}{a character vector}
    \item{\code{species_cn}}{a character vector}
    \item{\code{species}}{a character vector}
    \item{\code{iucn_red_list_category}}{a character vector}
    \item{\code{iucn_red_list_criteria}}{a character vector}
    \item{\code{endemic_to_china}}{a character vector}
    \item{\code{source}}{a character vector}
  }
}
\details{

For more details about the categories, see https://www.iucnredlist.org/resources/categories-and-criteria

}
\source{
https://www.mee.gov.cn/xxgk2018/xxgk/xxgk01/202305/t20230522_1030745.html
}

\references{

China Biodiversity Red List (2020) higher plants, released by The Ministry of Ecology and Environment and The Chinese Academy of Sciences, The People's Republic of China on 19 May 2023.

https://www.mee.gov.cn/xxgk2018/xxgk/xxgk01/202305/t20230522_1030745.html

}

\examples{
data(dat_CBRL2020_higher_plants)

}
\keyword{datasets}
