% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{format_output}
\alias{format_output}
\title{Select and reorder and column names for the dataset}
\usage{
format_output(x)
}
\arguments{
\item{x}{a dataframe derived from List of plant species in China with more columns}
}
\value{
a reordered List of plant species in China
}
\description{
Select and reorder and column names for the dataset
}
\examples{

temp <- head(LPSC::LPSC2022)
temp2 <- cbind(temp, YOUR_SEARCH = paste0("something_entered", 1:nrow(temp)))
format_output(temp2)

}
