% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{get_accepted_name}
\alias{get_accepted_name}
\title{Search the accepted name for a scientific name or a Chinese name in List of plant species in China}
\usage{
get_accepted_name(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character string containing the species' scientific name or Chinese name}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the accepted name for a scientific name or a Chinese name in List of plant species in China
}
\examples{

aaaa <- get_accepted_name(c("小叶青冈",
"Cyclobalanopsis myrsinifolia",
"Machilus chekiangensis",
"宫廷胡萝卜"))
make_taxa_table(aaaa)
get_accepted_name(c("Carex phaeopoda"))
get_accepted_name(c("Cyclobalanopsis myrsinifolia"))
get_accepted_name(c("Carex phaeopoda2",
"Carex phaeopoda"))
get_accepted_name(c("Carex phaeopoda",
"苹果", "华中山楂", "华中山楂2"))
# 直接查询出接受名，中文，拉丁文均可
get_accepted_name("华中山楂2")
xxx <- get_accepted_name(c("Carex phaeopoda",
 "苹果", "华中山楂"))
make_taxa_table(xxx)
get_accepted_name("刺枝麻黄")
ttt <- get_accepted_name("单子麻黄2")
make_taxa_table(ttt)



}
