% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{get_accepted_name_one}
\alias{get_accepted_name_one}
\title{Search the accepted name for a scientific name in List of plant species in China}
\usage{
get_accepted_name_one(x0, y0 = LPSC::LPSC2022)
}
\arguments{
\item{x0}{a character string containing one scientific name only}

\item{y0}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the accepted name for a scientific name in List of plant species in China
}
\examples{
get_accepted_name_one(c("Carex phaeopoda"))
get_accepted_name_one(c("Cyclobalanopsis myrsinifolia"))
get_accepted_name_one("Adiantum pedatum  ")
get_accepted_name_one("Adiantum pedatum2  ")
get_accepted_name_one("Machilus chekiangensis")
get_accepted_name_one("Quercus myrsinifolia")
get_accepted_name_one("Cyclobalanopsis myrsinifolia")


}
