% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{REPLACE}
\alias{REPLACE}
\title{replace multiple white space}
\usage{
REPLACE(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string whose leading white space multiple white spaces have been replaced.
}
\description{
replace multiple white space
}
\examples{

REPLACE("   Michelia    alba   ")


}
