% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_canonical_name}
\alias{search_canonical_name}
\title{Search the List of plant species in China using the canonical name}
\usage{
search_canonical_name(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character vector, each element represents a scientific name}

\item{dat}{the dataset used, here the dataframe of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the List of plant species in China using the canonical name
}
\examples{

search_canonical_name(x = "ginkgo biloba") # 全名匹配
search_canonical_name(x = "   ginkgo  biloba ") # 全名匹配
search_canonical_name(x = "Ginkgo biloba") # 全名匹配
search_canonical_name(x = "Ginkgo biloba ") # 全名匹配
search_canonical_name(x = "Ginkgo  biloba ") # 全名匹配
search_canonical_name(x = "Ginkgo biloba2") # 全名匹配
search_canonical_name(x = c("Ginkgo biloba2",
"Salix wangiana",
"Machilus chekiangensis",
"Ziziphus jujuba")) # 全名匹配
aaa <- c("Ginkgo biloba2",
"Salix wangiana",
"Machilus chekiangensis",
"Ziziphus jujuba")
search_canonical_name(aaa)


}
