% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_family}
\alias{search_family}
\title{Search the List of plant species in China using Family Names}
\usage{
search_family(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character string representing the Family Names}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the List of plant species in China using Family Names
}
\examples{

search_family("Poaceae") # 完整匹配

}
