% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_family_c}
\alias{search_family_c}
\title{Search the List of plant species in China using Chinese Family Names}
\usage{
search_family_c(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a string in Chinese}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the List of plant species in China using Chinese Family Names
}
\examples{
search_family_c("连香树科") # 完整匹配


}
