% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_genus}
\alias{search_genus}
\title{Search the List of plant species in China using genera}
\usage{
search_genus(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character vector containing the genera's names}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the List of plant species in China using genera
}
\examples{

search_genus("Ginkgo")
search_genus("michelia")
search_genus("Lindera")
search_genus(c("Ginkgo", "木兰"))
search_genus(c("Ginkgo", "木兰", "Machilus", "Acer"))


}
