% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_infraspecies_c}
\alias{search_infraspecies_c}
\title{Search the List of plant species in China using a infraspecies' Chinese name}
\usage{
search_infraspecies_c(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character string containing the infraspecies' Chinese name}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the List of plant species in China using a infraspecies' Chinese name
}
\examples{

search_infraspecies_c("刺枝麻黄")
search_infraspecies_c(c("刺枝麻黄", "云实"))

}
