% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_source}
\alias{search_source}
\title{Search the literature where the name was published}
\usage{
search_source(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character string representing the name of the literature}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search the literature where the name was published
}
\examples{

search_source("Sargent") # 部分匹配

}
