% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{search_species_c}
\alias{search_species_c}
\title{Search List of plant species in China using a species' Chinese name}
\usage{
search_species_c(x, dat = LPSC::LPSC2022)
}
\arguments{
\item{x}{a character string containing species' Chinese names}

\item{dat}{a dataframe containing all the data of the List of plant species in China}
}
\value{
a subset of dat, only the matched entries
}
\description{
Search List of plant species in China using a species' Chinese name
}
\examples{

search_species_c("银杏")
search_species_c(c("银杏",
"篦子三尖杉", "八角莲",
"扁果草", "黄樟", "黄樟2"))


}
