% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{show_detail}
\alias{show_detail}
\title{Show details of species}
\usage{
show_detail(x, db = LPSC::dat_all_accepted_sp2023)
}
\arguments{
\item{x}{a character string representing scientific names or Chinese names of plants}

\item{db}{database used}
}
\value{
a dataframe showing the details of accepted species of List of Plant Species China
}
\description{
Show details of species
}
\examples{

res1 <- show_detail(c("木栏","香港木兰"))
show_detail(c("樟叶泡花树", "Machilus"))
show_detail("Machilus gamblei")
show_detail(c("大叶木兰","香港木兰","雾水葛"))
show_detail("滇青冈")
show_detail(c("滇青冈", "Quercus glaucoides"))
show_detail(c("", "Quercus glaucoides"))
show_detail(c("", ""))

}
