% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{show_iucn_status}
\alias{show_iucn_status}
\title{Search for China Biodiversity Red List using species names}
\usage{
show_iucn_status(x, db = LPSC::dat_CBRL2020_higher_plants)
}
\arguments{
\item{x}{Character string}

\item{db}{databased used in the analysis, by default China Biodiversity Red List (2020) higher plants}
}
\value{
a dataframe containing the results
}
\description{
Search for China Biodiversity Red List using species names
}
\examples{
library(LPSC)
show_iucn_status("竹叶青冈")
show_iucn_status("木姜叶青冈")
show_iucn_status("Quercus glauca")
show_iucn_status(c("樟叶泡花树", "香港木兰", "蛇藤", "匙叶黄杨", "海南姜"))
show_iucn_status(c("樟叶泡花树", "香港木兰", "蛇藤", "", "Wikstroemia nutans"))

}
