% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_accepted_name}
\alias{get_accepted_name}
\title{Look up the accepted name for given scientific names}
\usage{
get_accepted_name(x, db = LPSC::dat_all_sp2023)
}
\arguments{
\item{x}{a character vector representing scientific names}

\item{db}{the database used, by default CPSC2023}
}
\value{
a dataframe showing the details of accepted names
}
\description{
Look up the accepted name for given scientific names
}
\examples{

get_accepted_name("Cinnamomum camphora")
get_accepted_name(c("Michelia maudiae", "Machilus", "Cinnamomum camphora"))
get_accepted_name(c("Michelia maudiae", "Cinnamomum camphora", "machilus", ""))
get_accepted_name("Michelia")
get_accepted_name(c("Michelia", "magnolia"))
get_accepted_name("Michelia maudiae")
get_accepted_name("Ziziphus jujuba")



}
